refactor existing ImagePicker to AttachmentPicker
